requirements.txt
# Django & REST Framework
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.27.2

# Database
psycopg2-binary==2.9.10

# Environment & Configuration
python-dotenv==1.0.1

# Production Server
gunicorn==23.0.0

# Development & Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==5.0.0
# django-debug-toolbar==4.5.0
ipdb==0.13.13

# Code Quality & Formatting
black==24.10.0
flake8==7.1.1
isort==5.13.2

# Security & Authentication (добавим для будущего использования)
python-decouple==3.8
django-environ==0.11.2

cryptography==44.0.0

django-storages[boto3]         # Последняя ~1.14.x + boto3 для S3/MinIO
boto3>=1.42.0
pillow==10.2.0